Setting Langserve with groq model/requirements.txt
langchain
langchain-core
python-dotenv
langchain_groq
fastapi
uvicorn
langserve
sse_starlette